+++ plotly/src/traces/scatter.rs
@@ -10,8 +10,8 @@ use crate::ndarray::ArrayTraces;
 use crate::{
     color::Color,
     common::{
-        Calendar, Dim, ErrorData, Fill, Font, HoverInfo, HoverOn, Label, Line, Marker, Mode,
-        Orientation, PlotType, Position, Visible,
+        Calendar, Dim, ErrorData, Fill, Font, HoverInfo, HoverOn, Label, LegendGroupTitle, Line,
+        Marker, Mode, Orientation, PlotType, Position, Visible,
     },
     private::{NumOrString, NumOrStringCollection},
     Trace,
@@ -73,6 +73,9 @@ where
     /// same time when toggling legend items.
     #[serde(rename = "legendgroup")]
     legend_group: Option<String>,
+    /// Set and style the title to appear for the legend group
+    #[serde(rename = "legendgrouptitle")]
+    legend_group_title: Option<LegendGroupTitle>,
     /// Sets the opacity of the trace.
     opacity: Option<f64>,
     /// Determines the drawing mode for this scatter trace. If the provided `Mode` includes
@@ -417,6 +420,7 @@ mod tests {
             .hover_template_array(vec!["hover_template"])
             .ids(vec!["1"])
             .legend_group("legend_group")
+            .legend_group_title(LegendGroupTitle::new("Legend Group Title"))
             .line(Line::new())
             .marker(Marker::new())
             .meta("meta")
@@ -464,6 +468,7 @@ mod tests {
             "hovertemplate": ["hover_template"],
             "ids": ["1"],
             "legendgroup": "legend_group",
+            "legendgrouptitle": {"text": "Legend Group Title"},
             "line": {},
             "marker": {},
             "meta": "meta",